webserver.py: Start timer thread after code compilation for example projects
For example projects - Display humidity and temperature, Tweeting
doorbell, the time needed to compile and flash the arduino is about ~6
seconds which is not accounted for when the timer starts running. Thus,
the timer runs only for timeout-6 seconds. This patch waits for code
compilation to complete in both these examples  after which the timer is
started so that the example runs for intended timeout period.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>