touch_sensor.html: Update the touch and relay sensor example Image
Ultra96-zynqmp bitstream runs MIPI design that uses GPIO E-L pins. As a
result, these pins cannot be used by input sensors such as the touch
sensor in touch and relay sensor example. In this sensor mezzanine
example, the touch sensor is being connected on GPIO pin AB instead of
G3. Hence the image with the sensor connections needs to be updated
along with the description

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>